[react-hw11] Moved smart components into "modules"
