--- react-2021-09/game-of-life/src/components/AppStateManager/AppStateManager.stories.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { AppStateManager } from './AppStateManager';
-
-export default {
-    title: 'AppStateManager',
-};
-
-// export const Game = () => {
-//     return <AppStateManager />;
-// };
