+++ react-2021-09/game-of-life/src/modules/AppStateManager/index.tsx
